John had moved one of the wrapper divs intended to go around all slots in an area down into the code for each individual slot, breaking the add slot button.
I fixed that and then commented _area.php to help John understand what
all of the divs are for. I agree they are numerous but there's a purpose
to each one which needs to be clearly understood before changes are made.

We could remove a lot of them when not logged in but that would make
it harder to style slots once and be done for both editing and non-editing
modes because it wouldn't be possible to assume a particular set of 
ancestor divs.



git-svn-id: http://svn.symfony-project.com/plugins/pkContextCMSPlugin/trunk@16914 ee427ae8-e902-0410-961c-c3ed070cd9f9