Interactive command line installation script. Can create an entire Symfony project with the CMS center-stage, or just add the plugins to an existing Symfony project. Checks for an appropriate Symfony version and makes sure you have netpbm.
New sample configuration files.

_login.php no longer contains hardcoded login/logout route names, which cause problems 
in a typical scenario in which routing has been overridden in the most straightforward way
to put the CMS front and center.

Fixed issues with admin privileges leaking in non-Trinity-like setups due to the fact that
some versions of sfDoctrineGuard don't purge credentials when a user times out.



git-svn-id: http://svn.symfony-project.com/plugins/pkContextCMSPlugin/trunk@19132 ee427ae8-e902-0410-961c-c3ed070cd9f9