The breadcrumb, subnav and tabs now run much faster and use less memory. This is implemented via new model class methods that return simple associative arrays of information about the peer pages in question rather than fully hydrating expensive Doctrine page and slot objects just to give you the slug, title, etc. If you have overridden _breadcrumb.php or _tabs.php you'll need to make some changes, but it's straightforward, see the plugin versions of those files.
The queries may be optimized further at a later date to do most of them as a single query, but the API won't need to change 
again for that.

The new philosophy here: why are you hydrating a page other than the current page? You probably shouldn't be doing that.

The accelerated queries are run at the PDO level since there is no need for the performance overhead of running them
through Doctrine to get this simple information needed on every page render.



git-svn-id: http://svn.symfony-project.com/plugins/pkContextCMSPlugin/trunk@21150 ee427ae8-e902-0410-961c-c3ed070cd9f9