Search is now included in the cmstest project.
Search engine updates, by default, no longer require a cron job.
They happen synchronously instead for simplicity. There is a 
performance price but it is not large in practice and this makes it
much easier to set up the CMS. You can still enable the old behavior
(see the README).



git-svn-id: http://svn.symfony-project.com/plugins/pkContextCMSPlugin/trunk@22515 ee427ae8-e902-0410-961c-c3ed070cd9f9